Структуры моделей вынесены в отдельные файлы. Добавлена "фабрика вопросов"
