+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,12 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+
+import Foundation
+
+struct QuizResultsViewModel {
+    let title: String
+    let text: String
+    let buttonText: String
+}
