+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,12 @@
+//
+//  QuizStepViewModel.swift
+//  MovieQuiz
+//
+
+import UIKit
+
+struct QuizStepViewModel {
+    let image: UIImage
+    let question: String
+    let questionNumber: String
+}
